Completed Level 3
The game can now be played to a point where all 3 powerups are obtained!

There's a lot of lag when shooting every single breakable wall, due to physics calculations - A possible fix is using sphere colliders instead of mesh colliders.

The fourth level needs to completed, then implemented in the main scene.

After that, it's a matter of LIGHTING IT UP, tidying up, having proper keybindings, and some simple triggers to start/end the game. Then actually making the .exe, which I have never done before.